After reviewing the routes, deleteemployeeshift should be in the userservice and controller, not the shiftrotation service and controller
