+++ server/schedulism/src/main/java/com/madamepapier/schedulism/controllers/ShiftRotationController.java
@@ -97,18 +97,4 @@ public class ShiftRotationController {
         }
     }
 
-    //Delete users shift, HR only
-    @DeleteMapping("{requesterId}/deleteSpecificShift/{userId}/{shiftId}")
-    public ResponseEntity<Void>deleteSpecificShift(@PathVariable long requesterId,
-                                                   @PathVariable Long userId,
-                                                   @PathVariable Long shiftId){
-        try{
-            shiftRotationService.deleteSpecificShifts(requesterId,userId,shiftId);
-           return new ResponseEntity<>(HttpStatus.OK);
-        } catch (ErrorResponseException e){
-            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
-        }
-    }
-
-
 }
