+++ server/schedulism/src/main/java/com/madamepapier/schedulism/services/ShiftRotationService.java
@@ -120,40 +120,5 @@ public class ShiftRotationService {
         }
         return shiftRotationRepository.findAllByHasBeenRequestedTrue();
     }
-
-//    Delete employee shifts
-public void deleteSpecificShifts (long requesterId,long userId, long shiftId){
-    User requester = userRepository.findById(requesterId)
-            .orElseThrow(()-> new ErrorResponseException(HttpStatus.FORBIDDEN));
-
-    if (requester.getUserRole() != UserRole.HR_EMPLOYEE) {
-        throw new CustomException("Only HR employees can delete shifts.");
-    }
-
-    User userToDeleteShift = userRepository.findById(userId)
-            .orElseThrow(()-> new ErrorResponseException(HttpStatus.NOT_FOUND));
-
-    List<ShiftRotation> shiftRotations = userToDeleteShift.getShiftRotations();
-    ShiftRotation shiftRotationToRemove= null;
-
-     for(ShiftRotation shiftRotation: shiftRotations){
-          if(shiftRotation.getId().equals(shiftId)){
-              shiftRotationToRemove = shiftRotation;
-              break;
-           }
-     }
-     if (shiftRotationToRemove != null){
-         shiftRotationRepository.deleteById(shiftRotationToRemove.getId());
-         shiftRotations.remove(shiftRotationToRemove);
-         userToDeleteShift.setShiftRotations(shiftRotations);
-         userRepository.save(userToDeleteShift);
-     }
-    }
 }
 
-
-//Pseudo code
-//I want to delete employee shift.
-//first i would need to check whether the person requesting to delete the individuals shift is an HR
-// I need to get the list of shift rotatiion fo the user
-//I will then need to insert the id of the user to delete and the specific shift ID
